requirements.txt
arabic-reshaper==2.1.4
asgiref==3.8.1
asn1crypto==1.5.1
autopep8==1.7.0
backports.zoneinfo==0.2.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
cryptography==42.0.8
cssselect2==0.7.0
distro==1.9.0
Django==4.2.13
django-cors-headers==3.13.0
django-environ==0.8.1
django-filter==24.2
django-rest-knox==4.2.0
django-sequences==2.6
django-user-agents==0.4.0
djangorestframework==3.15.2
et-xmlfile==1.1.0
future==0.18.2
html5lib==1.1
httplib2==0.22.0
idna==3.4
Levenshtein==0.20.5
lxml==4.9.1
numpy==1.24.4
openpyxl==3.1.5
oscrypto==1.3.0
pandas==2.0.3
pdfkit==1.0.0
Pillow==9.2.0
psycopg2-binary==2.9.3
pycodestyle==2.9.1
pycparser==2.21
pyHanko==0.15.0
pyhanko-certvalidator==0.19.5
pyparsing==3.1.2
PyPDF3==1.0.6
python-bidi==0.4.2
python-dateutil==2.9.0.post0
python-decouple==3.7
python-http-client==3.3.7
python-Levenshtein==0.20.5
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
rapidfuzz==2.15.2
reportlab==3.6.11
requests==2.28.1
six==1.16.0
sqlparse==0.5.0
svglib==1.4.1
tabula-py==2.3.0
tinycss2==1.2.1
toml==0.10.2
tqdm==4.64.1
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==4.2
ua-parser==0.18.0
uritools==4.0.0
urllib3==1.26.12
user-agents==2.2.0
webencodings==0.5.1
xhtml2pdf==0.2.8
xlwt==1.3.0

